Added Student File for Project Level 3
This is the hardest version of the projects that we are doing usually there will be 4 but for this project there are only 3. Please read the word document that is attached and follow the instructions. There is an EXACT right answer to these so please follow the rules for syntax and coding. Good Luck.